+++ music_app/static/js/runtime-bundle.js
@@ -2056,7 +2056,12 @@ function parseCurrentBrowserUrlState() {
 }
 
 function pushBrowserViewState(view, stateSnapshot = view) {
-  window.history.pushState(stateSnapshot, '', buildUrl(view));
+  const settingsNavigation = window.AlbumHavenSettingsNavigation?.instance;
+  if (settingsNavigation?.pushLibraryHistory) {
+    settingsNavigation.pushLibraryHistory(buildUrl(view), stateSnapshot);
+  } else {
+    window.history.pushState(stateSnapshot, '', buildUrl(view));
+  }
 }
 
 // END js/runtime/browser-navigation-helpers.js